# specs/002-improve-wellmio-booking/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Wellmio Booking API
  version: 1.0.0
paths:
...
  /api/admin/timeslots:
    post:
      summary: Create a new timeslot (for admins)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeSlot'
      responses:
        '201':
          description: The created timeslot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeSlot'
  /api/admin/timeslots/{id}:
    put:
      summary: Update a timeslot (for admins)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeSlot'
      responses:
        '200':
          description: The updated timeslot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeSlot'
    delete:
      summary: Delete a timeslot (for admins)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No content
...
components:
  schemas:
    TimeSlot:
      type: object
      properties:
        id:
          type: string
          format: uuid
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        is_booked:
          type: boolean
...